modification 3 src/BmiService.java

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
         BmiService service = new BmiService();
-        int Bmi = (int) service.calculate(1.87, 98);
+        int Bmi = service.calculate(1.87, 98);
         System.out.println(Bmi);
     }
 }
